Supervisory Mode
Introduce Supervisory Mode for events - 40% limit on driver controller when DIO3 = TRUE. In event of override on operator controller (click on LEFT JOYSTICK), the operator control inherits full control of the drivetrain (returning power to 100%), and the driver controller is essentially nullified. To release control back to the driver controller, click the RIGHT JOYSTICK. The DIO3 switch must be in the true position, or these functions will not work.